refactor: move runtime proxy from internal/ to pkg/
Go's internal/ rule blocks cross-module imports of internal/runtime/proxy
from external consumers (e.g. clawvisor-cloud, where the runtime proxy
will be wired into a managed cloud service). Move the package to
pkg/runtime/proxy with byte-identical exported surface so external
consumers can import it directly.

Sibling packages under internal/runtime/ (autovault, conversation,
leases, policy, review, timing) stay in internal/ — they remain
importable from pkg/runtime/proxy (same module) but are not exposed to
other modules. No sibling-package types appear in pkg/runtime/proxy's
exported API.

Caller imports updated:
- pkg/clawvisor/run.go
- internal/api/handlers/runtime.go and runtime_test.go
- cmd/clawvisor/cmd_agent_runtime.go
- docs/runtime-capability-matrix.md (doc references)

No API or behavior changes. Verified: go vet, go build, go test ./...
all pass.
